addded prompts

+++ appconfig/tmuxinator/install.sh
@@ -1,10 +1,23 @@
 #!/bin/bash
 
-# install tmuxinator
-cd $SUBMOBULES_PATH/tmuxinator
-git pull
-sudo gem install tmuxinator
+# get the path to this script
+APP_PATH=`dirname "$0"`
+APP_PATH=`( cd "$APP_PATH" && pwd )`
 
-# symlink tmuxinator settings
-rm ~/.tmuxinator
-ln -s $APPCONFIG_PATH/tmuxinator/dottmuxinator ~/.tmuxinator
+read -r -p $'\033[31mInstall TMUXINATOR? [y/n] \033[00m' response
+
+response=${response,,} # tolower
+if [[ $response =~ ^(yes|y| ) ]]; then
+
+  toilet Installing tmuxinator
+
+  # install tmuxinator
+  cd $APP_PATH/../../submodules/tmuxinator
+  git pull
+  sudo gem install tmuxinator
+  
+  # symlink tmuxinator settings
+  rm ~/.tmuxinator
+  ln -s $APP_PATH/dottmuxinator ~/.tmuxinator
+  
+fi
